+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@valora/logging",
   "version": "1.0.0",
-  "main": "dist/log.js",
-  "typings": "dist/log.d.ts",
+  "main": "dist/logging.js",
+  "typings": "dist/logging.d.ts",
   "repository": "https://github.com/valora-inc/logging",
   "author": "Valora Inc",
   "license": "Apache-2.0",
